examples/backend_service/requirements.txt
# Backend monitoring sidecar service dependencies
#
# Python version compatibility:
#   Python 3.8+  : use flask>=2.3,<3.1 and psutil>=5.9
#   Python 3.6/3.7: use the versions below (flask 2.0.x is the last series supporting Python 3.6)
#
# These versions are tested with Python 3.6.8+

flask==2.0.3
psutil==5.9.5
requests==2.28.2
python-ldap==3.4.4  # Optional — remove if LDAP group check endpoint not needed
